chore(analytics): export isEnabled directly
