--- shared/analytics/src/internals/enabled.ts
@@ -1,5 +1,0 @@
-import { getEnabledFn } from "../registry";
-
-export function isEnabled(): boolean {
-  return getEnabledFn()?.() ?? false;
-}
